Use syscall when name_to_handle_at is not available
This makes cgroupid function to work fine when building with alpine.
